Rename yaml validation files
