+++ clowder/clowder/yaml/validation/defaults.py
